[Fusili] Use pre-built IREE rather than pip dependency (#1907)
When generating a .vmfb we'll need to store it in a type that IREE
runtime expects; I believe that's `iree_const_byte_span_t`. We'll need
IREE cmake config + headers to able to use said type in sharkfuser. In
preparation for .vmfb generation, this PR moves towards pulling in IREE
tools in away that allows us to use cmake + headers + pre-built `.so`s.