[Fusilli] batch profiling of commands to automate data collection (#2590)
See: https://github.com/nod-ai/shark-ai/issues/2479

This is a relatively simple script as of right now, looking for feedback
as to what other stats to grab from the rocprof csv's, also there are
other rocprof args we can get stats from as well. These should be
relatively easy to add, though. Also, note that unlike the commands
pasted in the original issue, the commands for Fusilli are a little
different (learned this the hard way). We don't do any special parsing.
I reworked this to be closer to the Boo driver. I am manually verifying
that the stat collection is correct.

---------

Signed-off-by: raayandhar <rdhar@amd.com>